requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
openpyxl==3.1.2
mistralai==1.0.0
python-dotenv==1.0.0
aiofiles==23.2.1

